playbooks/team-sync/reference-impl/requirements.txt
fastapi==0.115.12
httpx==0.28.1
python-dotenv==1.0.1
requests==2.32.3
uvicorn[standard]==0.34.0
